Moved all remaining DDL tests to sqle package. Now need to make them pass.
