+++ go/libraries/doltcore/sqle/sqlddl_test.go
@@ -16,6 +16,8 @@ package sqle
 
 import (
 	"context"
+
+	"github.com/liquidata-inc/dolt/go/libraries/doltcore/row"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -336,3 +338,451 @@ func TestDropTable(t *testing.T) {
 		})
 	}
 }
+
+func TestAddColumn(t *testing.T) {
+	tests := []struct {
+		name           string
+		query          string
+		expectedSchema schema.Schema
+		expectedRows   []row.Row
+		expectedErr    string
+	}{
+		{
+			name:  "alter add column not null",
+			query: "alter table people add (newColumn varchar(80) not null default 'default' comment 'tag:100')",
+			expectedSchema: dtestutils.AddColumnToSchema(PeopleTestSchema,
+				schema.NewColumn("newColumn", 100, types.StringKind, false, schema.NotNullConstraint{})),
+			expectedRows: dtestutils.AddColToRows(t, AllPeopleRows, 100, types.String("default")),
+		},
+		{
+			name:  "alter add column not null with expression default",
+			query: "alter table people add (newColumn int not null default 2+2/2 comment 'tag:100')",
+			expectedSchema: dtestutils.AddColumnToSchema(PeopleTestSchema,
+				schema.NewColumn("newColumn", 100, types.IntKind, false, schema.NotNullConstraint{})),
+			expectedRows: dtestutils.AddColToRows(t, AllPeopleRows, 100, types.Int(3)),
+		},
+		{
+			name:  "alter add column not null with negative expression",
+			query: "alter table people add (newColumn float not null default -1.1 comment 'tag:100')",
+			expectedSchema: dtestutils.AddColumnToSchema(PeopleTestSchema,
+				schema.NewColumn("newColumn", 100, types.FloatKind, false, schema.NotNullConstraint{})),
+			expectedRows: dtestutils.AddColToRows(t, AllPeopleRows, 100, types.Float(-1.1)),
+		},
+		{
+			name:        "alter add column not null with type mismatch in default",
+			query:       "alter table people add (newColumn float default 'not a number' comment 'tag:100')",
+			expectedErr: "Type mismatch",
+		},
+		{
+			name:        "alter add column with tag conflict",
+			query:       "alter table people add (newColumn float default 1.0 comment 'tag:1')",
+			expectedErr: "A column with the tag 1 already exists",
+		},
+		{
+			name:        "alter add column not null without default",
+			query:       "alter table people add (newColumn varchar(80) not null comment 'tag:100')",
+			expectedErr: "a default value must be provided",
+		},
+		{
+			name:  "alter add column nullable",
+			query: "alter table people add (newColumn bigint comment 'tag:100')",
+			expectedSchema: dtestutils.AddColumnToSchema(PeopleTestSchema,
+				schema.NewColumn("newColumn", 100, types.IntKind, false)),
+			expectedRows: AllPeopleRows,
+		},
+		{
+			name:  "alter add column with optional column keyword",
+			query: "alter table people add column (newColumn varchar(80) comment 'tag:100')",
+			expectedSchema: dtestutils.AddColumnToSchema(PeopleTestSchema,
+				schema.NewColumn("newColumn", 100, types.StringKind, false)),
+			expectedRows: AllPeopleRows,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dEnv := dtestutils.CreateTestEnv()
+			CreateTestDatabase(dEnv, t)
+			ctx := context.Background()
+			root, _ := dEnv.WorkingRoot(ctx)
+
+			updatedRoot, err := ExecuteSql(dEnv, root, tt.query)
+
+			if tt.expectedErr == "" {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.expectedErr)
+				return
+			}
+
+			assert.NotNil(t, updatedRoot)
+			table, _, err := updatedRoot.GetTable(ctx, PeopleTableName)
+			assert.NoError(t, err)
+			sch, err := table.GetSchema(ctx)
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedSchema, sch)
+
+			updatedTable, ok, err := updatedRoot.GetTable(ctx, "people")
+			assert.NoError(t, err)
+			require.True(t, ok)
+
+			rowData, err := updatedTable.GetRowData(ctx)
+			assert.NoError(t, err)
+			var foundRows []row.Row
+			err = rowData.Iter(ctx, func(key, value types.Value) (stop bool, err error) {
+				r, err := row.FromNoms(tt.expectedSchema, key.(types.Tuple), value.(types.Tuple))
+				assert.NoError(t, err)
+				foundRows = append(foundRows, r)
+				return false, nil
+			})
+
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedRows, foundRows)
+		})
+	}
+}
+
+func TestUnsupportedAlterStatements(t *testing.T) {
+	tests := []struct {
+		name        string
+		query       string
+		expectedErr string
+	}{
+		{
+			name:        "alter add index",
+			query:       "alter table people add index myidx on (id, first)",
+			expectedErr: "Unsupported",
+		},
+		{
+			name:        "create index",
+			query:       "create index myidx on people (id, first)",
+			expectedErr: "Unsupported",
+		},
+		{
+			name:        "alter drop index",
+			query:       "alter table people drop index myidx",
+			expectedErr: "Unsupported",
+		},
+		{
+			name:        "drop index",
+			query:       "drop index myidx on people",
+			expectedErr: "Unsupported",
+		},
+		{
+			name:        "alter change column",
+			query:       "alter table people change id newId (varchar(80) not null)",
+			expectedErr: "Unsupported",
+		},
+		{
+			name:        "alter add foreign key",
+			query:       "alter table appearances add constraint people_id_ref foreign key (id) references people (id)",
+			expectedErr: "Unsupported",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dEnv := dtestutils.CreateTestEnv()
+			CreateTestDatabase(dEnv, t)
+			ctx := context.Background()
+			root, _ := dEnv.WorkingRoot(ctx)
+
+			_, err := ExecuteSql(dEnv, root, tt.query)
+			require.Error(t, err)
+			assert.Contains(t, err.Error(), tt.expectedErr)
+		})
+	}
+}
+
+func TestDropColumn(t *testing.T) {
+	tests := []struct {
+		name           string
+		query          string
+		expectedSchema schema.Schema
+		expectedRows   []row.Row
+		expectedErr    string
+	}{
+		{
+			name:           "alter drop column",
+			query:          "alter table people drop rating",
+			expectedSchema: dtestutils.RemoveColumnFromSchema(PeopleTestSchema, RatingTag),
+			expectedRows:   dtestutils.ConvertToSchema(dtestutils.RemoveColumnFromSchema(PeopleTestSchema, RatingTag), AllPeopleRows...),
+		},
+		{
+			name:           "alter drop column with optional column keyword",
+			query:          "alter table people drop column rating",
+			expectedSchema: dtestutils.RemoveColumnFromSchema(PeopleTestSchema, RatingTag),
+			expectedRows:   dtestutils.ConvertToSchema(dtestutils.RemoveColumnFromSchema(PeopleTestSchema, RatingTag), AllPeopleRows...),
+		},
+		{
+			name:        "drop primary key",
+			query:       "alter table people drop column id",
+			expectedErr: "Cannot drop column in primary key",
+		},
+		{
+			name:        "table not found",
+			query:       "alter table notFound drop column id",
+			expectedErr: "Unknown table: 'notFound'",
+		},
+		{
+			name:        "column not found",
+			query:       "alter table people drop column notFound",
+			expectedErr: "Unknown column: 'notFound'",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dEnv := dtestutils.CreateTestEnv()
+			CreateTestDatabase(dEnv, t)
+			ctx := context.Background()
+			root, _ := dEnv.WorkingRoot(ctx)
+
+			updatedRoot, err := ExecuteSql(dEnv, root, tt.query)
+
+			if tt.expectedErr == "" {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.expectedErr)
+				return
+			}
+
+			require.NotNil(t, updatedRoot)
+			table, _, err := updatedRoot.GetTable(ctx, PeopleTableName)
+			assert.NoError(t, err)
+			sch, err := table.GetSchema(ctx)
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedSchema, sch)
+
+			updatedTable, ok, err := updatedRoot.GetTable(ctx, "people")
+			assert.NoError(t, err)
+			require.True(t, ok)
+
+			rowData, err := updatedTable.GetRowData(ctx)
+			assert.NoError(t, err)
+			var foundRows []row.Row
+			err = rowData.Iter(ctx, func(key, value types.Value) (stop bool, err error) {
+				updatedSch, err := updatedTable.GetSchema(ctx)
+				assert.NoError(t, err)
+				r, err := row.FromNoms(updatedSch, key.(types.Tuple), value.(types.Tuple))
+				assert.NoError(t, err)
+				foundRows = append(foundRows, r)
+				return false, nil
+			})
+
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedRows, foundRows)
+		})
+	}
+}
+
+func TestRenameColumn(t *testing.T) {
+	tests := []struct {
+		name           string
+		query          string
+		expectedSchema schema.Schema
+		expectedRows   []row.Row
+		expectedErr    string
+	}{
+		{
+			name:  "alter rename column with column and as keywords",
+			query: "alter table people rename column rating as newRating",
+			expectedSchema: dtestutils.CreateSchema(
+				schema.NewColumn("id", IdTag, types.IntKind, true, schema.NotNullConstraint{}),
+				schema.NewColumn("first", FirstTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("last", LastTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("is_married", IsMarriedTag, types.BoolKind, false),
+				schema.NewColumn("age", AgeTag, types.IntKind, false),
+				schema.NewColumn("newRating", RatingTag, types.FloatKind, false),
+				schema.NewColumn("uuid", UuidTag, types.UUIDKind, false),
+				schema.NewColumn("num_episodes", NumEpisodesTag, types.UintKind, false),
+			),
+			expectedRows: AllPeopleRows,
+		},
+		{
+			name:  "alter rename column with column and to keyword",
+			query: "alter table people rename column rating to newRating",
+			expectedSchema: dtestutils.CreateSchema(
+				schema.NewColumn("id", IdTag, types.IntKind, true, schema.NotNullConstraint{}),
+				schema.NewColumn("first", FirstTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("last", LastTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("is_married", IsMarriedTag, types.BoolKind, false),
+				schema.NewColumn("age", AgeTag, types.IntKind, false),
+				schema.NewColumn("newRating", RatingTag, types.FloatKind, false),
+				schema.NewColumn("uuid", UuidTag, types.UUIDKind, false),
+				schema.NewColumn("num_episodes", NumEpisodesTag, types.UintKind, false),
+			),
+			expectedRows: AllPeopleRows,
+		},
+		{
+			name:  "alter rename primary key column",
+			query: "alter table people rename column id to newId",
+			expectedSchema: dtestutils.CreateSchema(
+				schema.NewColumn("newId", IdTag, types.IntKind, true, schema.NotNullConstraint{}),
+				schema.NewColumn("first", FirstTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("last", LastTag, types.StringKind, false, schema.NotNullConstraint{}),
+				schema.NewColumn("is_married", IsMarriedTag, types.BoolKind, false),
+				schema.NewColumn("age", AgeTag, types.IntKind, false),
+				schema.NewColumn("rating", RatingTag, types.FloatKind, false),
+				schema.NewColumn("uuid", UuidTag, types.UUIDKind, false),
+				schema.NewColumn("num_episodes", NumEpisodesTag, types.UintKind, false),
+			),
+			expectedRows: AllPeopleRows,
+		},
+		{
+			name:        "table not found",
+			query:       "alter table notFound rename column id to newId",
+			expectedErr: "Unknown table: 'notFound'",
+		},
+		{
+			name:        "column not found",
+			query:       "alter table people rename column notFound to newNotFound",
+			expectedErr: "Unknown column: 'notFound'",
+		},
+		{
+			name:        "column name collision",
+			query:       "alter table people rename column id to age",
+			expectedErr: "A column with the name 'age' already exists",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dEnv := dtestutils.CreateTestEnv()
+			CreateTestDatabase(dEnv, t)
+			ctx := context.Background()
+			root, _ := dEnv.WorkingRoot(ctx)
+
+			updatedRoot, err := ExecuteSql(dEnv, root, tt.query)
+
+			if tt.expectedErr == "" {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.expectedErr)
+				return
+			}
+
+			require.NotNil(t, updatedRoot)
+			table, _, err := updatedRoot.GetTable(ctx, PeopleTableName)
+			assert.NoError(t, err)
+			sch, err := table.GetSchema(ctx)
+			assert.Equal(t, tt.expectedSchema, sch)
+
+			updatedTable, ok, err := updatedRoot.GetTable(ctx, "people")
+			assert.NoError(t, err)
+			require.True(t, ok)
+
+			rowData, err := updatedTable.GetRowData(ctx)
+			assert.NoError(t, err)
+			var foundRows []row.Row
+			err = rowData.Iter(ctx, func(key, value types.Value) (stop bool, err error) {
+				updatedSch, err := updatedTable.GetSchema(ctx)
+				assert.NoError(t, err)
+				r, err := row.FromNoms(updatedSch, key.(types.Tuple), value.(types.Tuple))
+				assert.NoError(t, err)
+				foundRows = append(foundRows, r)
+				return false, nil
+			})
+
+			assert.NoError(t, err)
+			assert.Equal(t, tt.expectedRows, foundRows)
+		})
+	}
+}
+
+func TestRenameTable(t *testing.T) {
+	tests := []struct {
+		name           string
+		query          string
+		oldTableName   string
+		newTableName   string
+		expectedSchema schema.Schema
+		expectedRows   []row.Row
+		expectedErr    string
+	}{
+		{
+			name:           "alter rename table",
+			query:          "rename table people to newPeople",
+			oldTableName:   "people",
+			newTableName:   "newPeople",
+			expectedSchema: PeopleTestSchema,
+			expectedRows:   AllPeopleRows,
+		},
+		{
+			name:           "alter rename table with alter syntax",
+			query:          "alter table people rename to newPeople",
+			oldTableName:   "people",
+			newTableName:   "newPeople",
+			expectedSchema: PeopleTestSchema,
+			expectedRows:   AllPeopleRows,
+		},
+		{
+			name:           "rename multiple tables",
+			query:          "rename table people to newPeople, appearances to newAppearances",
+			oldTableName:   "appearances",
+			newTableName:   "newAppearances",
+			expectedSchema: AppearancesTestSchema,
+			expectedRows:   AllAppsRows,
+		},
+		{
+			name:        "table not found",
+			query:       "rename table notFound to newNowFound",
+			expectedErr: "Unknown table: 'notFound'",
+		},
+		{
+			name:        "table name in use",
+			query:       "rename table people to appearances",
+			expectedErr: "A table with the name 'appearances' already exists",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			dEnv := dtestutils.CreateTestEnv()
+			CreateTestDatabase(dEnv, t)
+			ctx := context.Background()
+			root, _ := dEnv.WorkingRoot(ctx)
+
+			updatedRoot, err := ExecuteSql(dEnv, root, tt.query)
+			if len(tt.expectedErr) > 0 {
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), tt.expectedErr)
+				return
+			} else {
+				require.NoError(t, err)
+			}
+			require.NotNil(t, updatedRoot)
+
+			has, err := updatedRoot.HasTable(ctx, tt.oldTableName)
+			require.NoError(t, err)
+			assert.False(t, has)
+			newTable, ok, err := updatedRoot.GetTable(ctx, tt.newTableName)
+			require.NoError(t, err)
+			require.True(t, ok)
+
+			sch, err := newTable.GetSchema(ctx)
+			require.NoError(t, err)
+			require.Equal(t, tt.expectedSchema, sch)
+
+			rowData, err := newTable.GetRowData(ctx)
+			require.NoError(t, err)
+			var foundRows []row.Row
+			err = rowData.Iter(ctx, func(key, value types.Value) (stop bool, err error) {
+				r, err := row.FromNoms(tt.expectedSchema, key.(types.Tuple), value.(types.Tuple))
+				require.NoError(t, err)
+				foundRows = append(foundRows, r)
+				return false, nil
+			})
+
+			require.NoError(t, err)
+
+			// Some test cases deal with rows declared in a different order than noms returns them, so use an order-
+			// insensitive comparison here.
+			assert.ElementsMatch(t, tt.expectedRows, foundRows)
+		})
+	}
+}
+
