+++ go/libraries/doltcore/sqle/testutil.go
@@ -18,11 +18,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	sqle "github.com/src-d/go-mysql-server"
+	"github.com/src-d/go-mysql-server/sql"
 	"io"
 	"strings"
 
-	sqle "github.com/src-d/go-mysql-server"
-	"github.com/src-d/go-mysql-server/sql"
 	"vitess.io/vitess/go/vt/sqlparser"
 
 	"github.com/liquidata-inc/dolt/go/libraries/doltcore/doltdb"
